Add base DeviceMeasurement class as a base for models in SDK
